+++ packages/gamut-styles/src/variables/deprecated-colors.ts
@@ -3,8 +3,12 @@ import { colors } from './colors';
 const { black, white } = colors;
 
 /**
- * @deprecated please use the `colors` export for the latest colors
+ * @deprecated
+ * Using these variables directly is no longer supported.
+ *
+ * Please use [`theme.colors`](https://gamut.codecademy.com/storybook/?path=/docs/foundations-theme--colors#standard-colors)
  */
+
 export const deprecatedColors = {
   blue: {
     '100': '#c8d7fa',
@@ -103,8 +107,11 @@ export const deprecatedColors = {
 
 /**
  * @deprecated
- * use the standard color set
+ * Using these variables directly is no longer supported.
+ *
+ * Please use [`theme.colors`](https://gamut.codecademy.com/storybook/?path=/docs/foundations-theme--colors#standard-colors)
  */
+
 export const brandColors = {
   red: deprecatedColors.red[500],
   orange: deprecatedColors.orange[400],