+++ packages/gamut/src/NotificationListNew/index.tsx
@@ -1,7 +1,7 @@
+import styled from '@emotion/styled';
 import { isEmpty } from 'lodash';
 import React from 'react';
 
-import { Box, Pattern } from '..';
 import { Notification } from '../NotificationList/typings';
 import { EmptyNotification } from './EmptyNotification';
 import { NotificationItemNew } from './NotificationItemNew';
@@ -10,36 +10,33 @@ export type NotificationListNewProps = {
   onDismiss?: (notification: Notification) => void;
   notifications: Notification[];
   onNotificationClick?: (notification: Notification) => void;
+  headerElementId?: string; // Used for aria-labelledby for the list.
 };
+
+const UnstyledUnorderedList = styled.ul`
+  list-style: none;
+  padding: 0;
+  margin: 0;
+`;
 
 export const NotificationListNew: React.FC<NotificationListNewProps> = ({
   notifications,
   onNotificationClick,
   onDismiss,
+  headerElementId,
 }) => {
-  const pattern = (
-    <Box paddingX={32}>
-      <Pattern name="dotsDense" height="1px" display="flex" />
-    </Box>
-  );
-  return (
-    <div>
-      {isEmpty(notifications) ? (
-        <EmptyNotification />
-      ) : (
-        notifications.map((notification: Notification) => {
-          return (
-            <React.Fragment key={notification.id}>
-              <NotificationItemNew
-                notification={notification}
-                handleClick={() => onNotificationClick?.(notification)}
-                handleDismiss={() => onDismiss?.(notification)}
-              />
-              {pattern}
-            </React.Fragment>
-          );
-        })
-      )}
-    </div>
+  return isEmpty(notifications) ? (
+    <EmptyNotification />
+  ) : (
+    <UnstyledUnorderedList aria-labelledby={headerElementId} aria-live="polite">
+      {notifications.map((notification: Notification) => (
+        <NotificationItemNew
+          key={notification.id}
+          notification={notification}
+          handleClick={() => onNotificationClick?.(notification)}
+          handleDismiss={() => onDismiss?.(notification)}
+        />
+      ))}
+    </UnstyledUnorderedList>
   );
 };