+++ packages/styleguide/stories/Layouts/Elements/index.tsx
@@ -1,5 +1,6 @@
 import { Box } from '@codecademy/gamut';
-import { colors } from '@codecademy/gamut-styles';
+// eslint-disable-next-line local-rules/gamut-import-paths
+import { colors } from '@codecademy/gamut-styles/src';
 import React from 'react';
 
 export const ExampleBox: React.FC = ({ children }) => (