+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.9.11](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.9.10...@codecademy/gamut-illustrations@0.9.11) (2021-04-12)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.9.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.9.9...@codecademy/gamut-illustrations@0.9.10) (2021-04-09)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.9.9](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.9.8...@codecademy/gamut-illustrations@0.9.9) (2021-04-08)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.9.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.9.7...@codecademy/gamut-illustrations@0.9.8) (2021-04-06)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
 ### [0.9.7](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.9.6...@codecademy/gamut-illustrations@0.9.7) (2021-04-05)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations