+++ packages/markdown-overrides/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.21](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.20...@codecademy/markdown-overrides@0.5.21) (2021-04-12)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.20](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.19...@codecademy/markdown-overrides@0.5.20) (2021-04-09)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.19](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.18...@codecademy/markdown-overrides@0.5.19) (2021-04-08)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.18](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.17...@codecademy/markdown-overrides@0.5.18) (2021-04-06)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
 ### [0.5.17](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.16...@codecademy/markdown-overrides@0.5.17) (2021-04-05)
 
 **Note:** Version bump only for package @codecademy/markdown-overrides